mtuci-summer-parser/requirements.txt
requests
python-dotenv
asyncpg
flask
flask-restx
flask-sqlalchemy
flask-cors